refactor: introduce onsager-github library; consolidate GitHub HTTP
Stops the bleeding from #221 (#220 Sub-issue A): GitHub HTTP was
scattered across stiglab + portal. Move the cleanly-extractable
surfaces into a new library-plane crate, treat GitHub as configurable
infra (NotConfigured = no-op), and put a CI lint behind the wall.

What moved into onsager-github:
- api::app          (App JWT, install token, list repos/labels, default branch)
- api::oauth        (authorize URL, code exchange, /user fetch)
- api::issues       (paginated issues, set_label, post_issue_comment)
- api::pulls        (paginated pulls, create_check_run + CheckConclusion)
- webhook::signature (HMAC-SHA256 verification — moved from portal)
- webhook::event    (WebhookEvent stub + to_spine_events placeholder;
                     full registry-backed translation lands with #150)
- credential        (Credential enum + GithubAuth trait + AccountKind)
- adapter::register (writes the artifact_adapters row at boot)
- error::GithubError (typed surface; NotConfigured is the feature-flag arm)

Wiring:
- stiglab + onsager-portal depend on onsager-github.
- stiglab/server/github_app.rs becomes a thin facade re-exporting the
  library items + mapping AccountKind ↔ core::GitHubAccountType for
  existing call sites.
- stiglab/server/auth.rs re-exports the OAuth helpers.
- adapter::register{,_any} called from stiglab + portal startup.

xtask lint-seams gains a github-http-wall check that forbids
api.github.com / github.com/{login/oauth,app/installations} literals
outside onsager-github. Two files (workflow_activation.rs, routes/
projects.rs) are still violators; they're listed in the explicit
GITHUB_HTTP_WALL_PENDING_FILES allowlist with reasons + follow-up
pointers, which CI prints loudly each run. They migrate in the
follow-up PR.

Deferred to follow-up (called out in PR description):
- Move webhook register/deregister + ensure_label from
  workflow_activation.rs (tangled with workflow URL resolution +
  ActivationError shape; needs careful surgery).
- Migrate stiglab/routes/projects.rs reads onto api::issues + api::pulls
  (lands with #220 Sub-issue B's portal split).
- Full WebhookEvent::to_spine_events translator (gated on #150).

No behavior change. 11 new unit tests in onsager-github (PEM normalize,
JWT signing, OAuth URL shape, signature verify) port over from the
moved files; existing stiglab + portal tests still pass.

Part of #220.
Closes #221.
