+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "crates/onsager-registry",
     "crates/onsager-spine",
     "crates/onsager",
+    "crates/onsager-github",
     "crates/onsager-portal",
     "crates/forge",
     "crates/ising",
