+++ crates/onsager-github/Cargo.toml
@@ -1,39 +1,32 @@
 [package]
-name = "onsager-portal"
+name = "onsager-github"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
-description = "GitHub webhook ingress for the Onsager factory — verifies, resolves, emits"
+description = "Typed GitHub client + webhook + OAuth for Onsager — the single home for GitHub HTTP"
 
 [lib]
-name = "onsager_portal"
+name = "onsager_github"
 path = "src/lib.rs"
-
-[[bin]]
-name = "onsager-portal"
-path = "src/main.rs"
 
 [dependencies]
 onsager-artifact = { path = "../onsager-artifact" }
-onsager-github = { path = "../onsager-github" }
 onsager-spine = { path = "../onsager-spine" }
-refract = { path = "../refract" }
 
 serde = { workspace = true }
 serde_json = { workspace = true }
 chrono = { workspace = true }
-uuid = { workspace = true }
-tokio = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-anyhow = { workspace = true }
 thiserror = { workspace = true }
+anyhow = { workspace = true }
 async-trait = { workspace = true }
-clap = { workspace = true }
-axum = { workspace = true }
+tracing = { workspace = true }
+
 reqwest = { workspace = true }
 sqlx = { workspace = true }
+jsonwebtoken = { workspace = true }
 ring = { workspace = true }
 hex = { workspace = true }
 base64 = { workspace = true }
-regex = "1"
+
+[dev-dependencies]
+tokio = { workspace = true }
