+++ crates/onsager-portal/src/lib.rs
@@ -30,9 +30,7 @@ pub mod backfill;
 pub mod config;
 pub mod db;
 pub mod gate;
-pub mod github;
 pub mod handlers;
 pub mod migrate;
 pub mod server;
-pub mod signature;
 pub mod state;
