+++ crates/onsager-portal/src/server.rs
@@ -15,6 +15,12 @@ pub async fn run(config: Config) -> anyhow::Result<()> {
     let pool = crate::db::connect(&config.database_url).await?;
     crate::migrate::run(&pool).await?;
     let spine = onsager_spine::EventStore::connect(&config.database_url).await?;
+    // Register the GitHub adapter into the spine catalog. Best-effort:
+    // a missing `artifact_adapters` table (older migration set) shouldn't
+    // block portal boot. See `onsager_github::adapter::register`.
+    if let Err(e) = onsager_github::adapter::register(&pool, "default").await {
+        tracing::warn!(error = %e, "github adapter registration skipped");
+    }
     let gate = Arc::new(GateClient::new(config.synodic_url.clone()));
 
     let state = AppState {
