+++ crates/stiglab/Cargo.toml
@@ -15,6 +15,7 @@ path = "src/main.rs"
 
 [dependencies]
 onsager-artifact = { path = "../onsager-artifact" }
+onsager-github = { path = "../onsager-github" }
 onsager-registry = { path = "../onsager-registry" }
 onsager-spine = { path = "../onsager-spine" }
